Refactored cross compilation tests
